separated the helper functions to their own file
